Improve terrain rendering performance
This is achieved by reducing the size of the edge/corner sprites: even though there are lots of empty pixels, the GPU calculates the triangles and checks whether the positions are empty or not. However, we now for certain that they are. This change reduces the sprite size and moves the sprites to their correct positions respectively. Therefore, some unnecessary GPU calculations are omitted.